fixing size_t

+++ testing/solutions/matrix-multiplication.cu
@@ -1,19 +1,19 @@
 #include <cuda_runtime.h>
 
 __global__ void matmul_kernel(const float* A, const float* B, float* C,
-                               int M, int N, int K) {
-    int row = blockIdx.y * blockDim.y + threadIdx.y;
-    int col = blockIdx.x * blockDim.x + threadIdx.x;
+                               size_t M, size_t N, size_t K) {
+    size_t row = blockIdx.y * blockDim.y + threadIdx.y;
+    size_t col = blockIdx.x * blockDim.x + threadIdx.x;
     if (row < M && col < N) {
         float sum = 0.0f;
-        for (int k = 0; k < K; k++)
+        for (size_t k = 0; k < K; k++)
             sum += A[row * K + k] * B[k * N + col];
         C[row * N + col] = sum;
     }
 }
 
 extern "C" void solution(const float* A, const float* B, float* C,
-                          int M, int N, int K) {
+                          size_t M, size_t N, size_t K) {
     dim3 block(16, 16);
     dim3 grid((N + 15) / 16, (M + 15) / 16);
     matmul_kernel<<<grid, block>>>(A, B, C, M, N, K);
